authorizationsForUser() takes str not Event (#1086)
* fix: correctly decode event ID (DO NOT SUBMIT)

There's some bug in here such that when I load IMS against SQLite,
I'm unable to retrieve saved Events. It's because those get read as
Event objects in this code location, rather than as strings.

This PR is super dirty, and it's here to notify Tool about the bug,
not really as a fix. I'm not even sure if it just afflicts the SQLite
scenario, but I haven't tried it with MySQL.

* Pass event ID, not event, to authorizationsForUser

* Shouldn't be necessary. Wish the type checker caught this.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: Wilfredo Sánchez <wsanchez@wsanchez.net>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>