feat: encourage user to log in with email address
This is a purely cosmetic change, since both email address and handle
currently work for logging into IMS.

I observed directly many people struggle to log in with their handles,
since they'd either get the capitalization wrong or they'd mess up
the presence of spaces/dashes.

People are already more comfortable with logging in with their email
addresses, since that's how they do it in the Clubhouse. This aligns
Clubhouse and IMS a bit better.

The "title" addition gives an alt-text style popup if you hover
over the "Email (or handle)" text.