feat: don't serialize personnel email to JSON
We have an authenticated API endpoint that lists all Rangers' handles,
names, and email addresses. That endpoint is currently only used to
serve the "Add Rangers" functionality on the incident page. Nothing
in the UI requires use of these email addresses. We can get a nice
little privacy win by not exposing them anymore. Note that this PR
uses the same approach to not serializing email that we already use
to not serialize passwords.

See here for an example of personnel:
https://ranger-ims.burningman.org/ims/api/personnel/

and note that no JS code references email:
https://github.com/search?q=repo%3Aburningmantech%2Franger-ims-server%20lang%3Ajs+email&type=code

and that the JS code in ranger-ims-web does expect the email field, but
it doesn't use it anywhere:
https://github.com/search?q=repo%3Aburningmantech%2Franger-ims-web%20lang%3Ajs+email&type=code