feat: allow more than 6 tabs per browser session
I was inspired by this comment on Reddit (the first in the list):
https://www.reddit.com/r/webdev/comments/149bjod/comment/kwm0742/

Rather than having each tab establish an eventsource connection
individually, this changes it so that all tabs share one eventsource.
That's done by having one tab become the leader (Web Locks API),
then having it fan out updates to all the other tabs in the session
(Broadcast Channel API).

If the leader tab's connection to the EventSource gets closed, then
the tab relinquishes the lock and allows another tab to set up the
connection again.