feat: allow more than 6 tabs per browser session :D
DO NOT MERGE yet, as this still has debug logging, and I'd like
to try it out a bit more first anyway.

I was inspired by this comment on Reddit (the first in the list):
https://www.reddit.com/r/webdev/comments/149bjod/comment/kwm0742/

Rather than having each tab establish an eventsource connection
invidually, this changes it so that all tabs share one eventsource.
That's done by having one tab become the leader (Web Locks API),
then having it fan out updates to all the other tabs in the session
(Broadcast Channel API).

I'm looking for feedback and alpha-testing at this point.