fix: don't throw DataTables errors
This seems to just cause the page to hang on "processing" when
I try to make an error in staging (rather than setting the error
text). It works locally though... I think the problem is that
we're throwing the error now and it's not being caught by
anything. Using "none" seems to work locally as well, so let's
try it out in staging.