fix: try again to show error message on incidents page
The change in https://github.com/burningmantech/ranger-ims-server/pull/1338
didn't suffice. I think this has a good chance though. I think in the AWS
case, we're calling the dataSrc handler with an empty set of data, calling
the line that clears the error text.

This tweak might fix it. It's annoying this behaves differently on errors
on AWS vs locally.