fix(.github): skip no-commit-to-branch on cicd workflow
This pre-commit is designed for local use, i.e. when you try to
commit to a specific branch locally. It doesn't make sense to run
this when actually trying to merge a commit to master via CICD,
since it'll always fail.

Note that this is commonly skipped in GitHub workflows:
https://sourcegraph.com/search?q=context:global+f:.github/workflows+content:no-commit-to-branch&patternType=regexp&sm=0