feat: use just two SQL calls to back the incidents API
Currently a call to the incidents API (which backs the incidents page)
results in over 5*n sequential DB lookups, where n is the number of
incidents in the specified event. Given that we get to over 1000
incidents per event, that starts to be quite a lot of queries. My
suspicion is that it'd be much faster to issue just a few big queries
instead, and that's what this change does.

On staging, this call takes about 10 seconds, as things are. I'm
hopeful that this change will bring that down significantly.
https://ranger-ims-staging.burningman.org/ims/api/events/2023/incidents/

I've confirmed locally for MySQL and for SQLite that this change
doesn't alter the API result for my bunch of test data.