fix "Add" formatting, use datalist and text for incident types
The "Add" textboxes were both artificially narrow. We might as
well use the flex-input-container class, which was already used
above on the page, for "Summary".

This also makes incident types a searchable text field, as I
recently made the case for the "Add Rangers" dropdown. As with
the Rangers list, this only permits values from the incident
types list.