stop shuffling ranger_handles and incident_types
Currently those fields under an incident get randomly sorted in
the API response, since a frozenset is unordered. This surfaces on
the incidents page, where ranger handles and incident types both
show up in a random order. That's kind of weird.

This is a pretty simple fix for that, which makes the API response
consistent.

I've looked through the Git history a fair bit to see why frozensets
were used in the first place. I'm fairly confident it was just
because we needed an immutable, hashable collection with duplicates
removed. This new way achieves those same ends too, but it also
maintains ordering.
https://github.com/search?q=repo%3Aburningmantech%2Franger-ims-server+frozenset&type=commits