fix "controlClear" function
A jQuery queue function needs to dequeue itself in order
to let subsequent functions run. As things were, for example,
if you modify the summary field, this function would correctly
clear the formatting after 1 second, but then on a second
modification, this function would not run and the formatting
would not unset.