upgrade mypy to support generics
I want to use PEP 695 generics in another PR, but mypy is blocking me:
"PEP 695 generics are not yet supported"

mypy-zope has to be upgraded too, because the old version of it is
incompatible with the newer mypy versions.

The various uv.lock changes are a result of pyproject.toml changes on
master that haven't propagated to uv.lock (our dependabot might need
some enhancement).