rename "incident reports" to "field reports" in UI
also start using the initialism "IMS #" for incident number
and "FR #" for field report number.

This is only a surface-level change, affecting pieces of IMS
that are visible to users. There's a bunch more work to do
sometime to rename everything else, from persistence to server
code to the APIs.

https://github.com/burningmantech/ranger-ims-server/issues/371#issuecomment-2460716389