tighten up spacing a bit to increase page density
Currently the incident page in particular is very spaced out,
so you can't see much on a single page. This densifies everything
a bit. It also removes the h1 title from the incident page, where
it provides no value, but takes up space.

Additionally this changes "Description" to "Details" for location,
so that the text is no longer so wide that it overflows onto the
textbox (yeah, lazy fix)