retry EventSource connection when it fails
I want to see if this resolves
https://github.com/burningmantech/ranger-ims-server/issues/1364

It seems kind of ugly having an infinite loop on errors,
but this is perhaps the only way to ensure that the client is
always trying to have an active EventSource connection. Even
on-playa where the 524 errors are unexpected, other transient
issues could cause the EventSource connections to sever, and
it'd probably be good for the clients to keep trying to
reconnect.