exclude system entries from calls for all incidents and incident reports
This does not change the default behavior of the APIs, as this adds a
new query param that must be set to "true" to get the new behavior.

This change should make the incidents page load noticeably faster,
since we'll put less pressure on the server, and the data retrieval
will drop from around 4 MB to 2 MB. Additionally, the search should
be faster, because we'll be searching over much less unnecessary text.

https://github.com/burningmantech/ranger-ims-server/issues/1375