remove time zone ambiguity, drop momentJS dependency
This makes it clear in the UI that times are all in the user's
time zone, not Pacific. It does this using modern JavaScript time
capabilities, so we can drop use of the long-deprecated momentJS
dependency.

I'd kind of like to extend this to always show times in LA time,
but it'd be kind of ugly to hardcode America/Los_Angeles in the
frontend code. This is a good stepping stone, anyway.

https://github.com/burningmantech/ranger-ims-server/issues/1407