migrate to Bootstrap 5, implement light/dark mode
The main reason to upgrade was to allow dark mode, which is something
various nighttime Operators have been requesting for a while. This
upgrade forced various other smaller changes too. The pages all look
a bit different, but hopefully mostly for the better.

I had to upgrade DataTables to get to a Bootstrap 5-compatible version.

One other tweak in this PR is that the instructions on the Field Report
page are turned into an accordion (a Bootstrap 5 thing), so that they
can be hidden by default.