don't send Set-Cookie back from eventsource calls
this fixes the below issue, in a minimally hacky way.
The eventsource call doesn't require authentication, and
when a client does automatic connection retries on this
endpoint, it'll sometimes get back a Set-Cookie value
for a brand new, unauthenticated Twisted session. That
cookie value can stomp over a valid, authenticated session
cookie that the client may already have. The result is that
a user can effectively be logged out after a session lasting
less than a minute. That's not desirable.

https://github.com/burningmantech/ranger-ims-server/issues/1363