only update the changed incident on background update
Previously on any update to any incident, we were pulling all the
incident data for every incident in the event. That would lead to
heavy server load when many users were logged into IMS at once, and
it make the table spin waiting for the large bundle of new data to
arrive.

https://github.com/burningmantech/ranger-ims-server/issues/1488