add some resiliency around missed SSEs
with my changes yesterday to only selectively update the Incidents page,
i.e. only update one row at a time, it's even more important that we
don't miss any SSEs. This change builds in some resiliency, by having
the Incidents and Incident pages do more substantial reloads when an
SSE is known to have been missed.

I'm not sure this is the final version of what I want in prod on playa,
but I'd like to try this out in staging, where the network connection
is less reliable for me than from localhost.