improve handling for potentially missed SSEs
This starts telling EventSource clients, upon connection, which SSE
ID was the most recently published. That allows the client to know
whether it should reload its data pages to be in sync with the server.
I store the previous SSE ID in LocalStorage, so that all tabs can share
that value together.

This makes the synchronization mechanism very resilient to the
EventSource connection needing to be reestablished.

This simplifies some of the _eventsource.py code by removing unused
code around replaying events (we were never using that).