appease ruff's new stdlib-module-shadowing (A005) rule
ruff wants to auto-update, but that isn't working, because it's now
angry that we have files named "json.py" and "enum.py". This commit
renames those two files, with large rename fallout.

Also in this commit, I tweak a Hypothesis return type to stop a
warning from showing up on every test run.