remove ** permissions from IMS, because they don't work
This feature didn't actually work as of now, because the incident
and field report pages required user authentication. That made this
** ACLing pointless, because it was all about granting access when
there was no authenticated user. We might as well just remove this
sort of ACL, rather than have it in here, but have it broken.

More details in attached screenshot in this PR. It would take some
different architecture to make this work.