don't show attached FRs for a new incident (#1560)
fixing a silly bug I introduced yesterday, in which a new incident was
showing as having a bunch of FRs attached to it. That's because I was
checking if "fr.incident === incidentNumber" for each of those FRs, and
it turns out that null == null, so we showed all those FRs that weren't
attached to any incident yet.