make TWISTED_SESSION cookie use more secure
Twisted doesn't set a SameSite value on its session cookie. Browsers are
supposed to interpret that as "lax", but Chrome interprets that as
something a little more lenient still than "lax". The right thing to do
is to actually set a SameSite value, and lax is appropriate for our
needs.
https://web.dev/articles/samesite-cookies-explained#samesitelax_by_default

HttpOnly forbids JavaScript from accessing the cookie directly (though
it'll still get passed implicitly when HTTP requests are made). This
helps protect against XSS, since a malicious script is unable to read a
cookie in order to steal it.
https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Set-Cookie#httponly

https://github.com/burningmantech/ranger-ims-server/issues/1562