permit regular expression searches in UI, allow bookmarking
I don't expect most users to notice this feature, which is totally fine
by me. The major use case for this is to allow "OR"-based searching,
which the current DataTables "smart search" approach doesn't support.

For example, this will let someone bookmark this (not URL-encoded here,
for clarity):
https://.../ims/app/events/2024/incidents/?q=/\b(dog|cat|snake)\b/
to easily find any incident that contains the word "dog" or "cat" or
"snake". The word boundary check might be useful too, as shown in that
example.

https://github.com/burningmantech/ranger-ims-server/issues/1570