replace URL query string on changes to search field
This makes bookmarking a lot more intuitive, since the URL
will always reflect the current state of the search field.
This approach works on the client-side to replace that URL,
using pushState.

After this, I plan to create query params for the other filters
on the Incidents and Field Reports pages, then add those values
to this updated URL as well.

https://github.com/burningmantech/ranger-ims-server/issues/1570