support multi-select of incident types on Incidents page
This also swaps us over to using the URL fragment (i.e. the part after
the "#") to save filter parameters. Those allow for longer inputs than
search params, and they're fully client-side, which is fine for our
purposes.

To support that, I had to get a bunch of onclick handlers to return
false, to prevent redirects to href="#", since that would stomp over
my attempts to set the fragment.