add "lastModified" to Incident entity, add column in DataTable (#1589)
Before this change, we were figuring out last modification time on the
client side. That was being calculated incorrectly, since we weren't
passing system report entries into the Incidents view
(exclude_system_entries = True). This moves over that computation to
the server.

Also, this adds a "Modified" column to the Incidents view. I'm
theorizing that people might find this useful; possibly more useful
than "Created". It could especially be useful as a sort order, so that
active Incidents are at the top and older ones fall to the bottom.

NOTE that if we want to remove this DataTables column in the future,
we should still keep the server-side piece of this commit, and we should
still use the new "last_modified" field in the API to know when an
Incident was last modified.

(finally, there's one little tweak in here for document titles, making
them more concise)