allow case insensitive login, add login tests (#1610)
The 2024 Tech AARs asked for case insensitive login, as we already have
with Clubhouse. I don't totally love this, because the specification
for SMTP clearly says
"The local-part of a mailbox MUST BE treated as case sensitive"
even though most mail servers treat it as case insensitive in practice.

My preference would be for us to only treat the part after the "@" as
case insensitive, but whatever we do, it should align with Clubhouse.

https://www.ietf.org/rfc/rfc2821.txt