require event IDs to match ^[\w_-]+$
Before this change, event IDs could be totally arbitrary strings.
That could make for ugly URLs. Plus, for the work I'm doing now
to support file attachments, I need to be able to use the event ID
as part of a local filesystem path. That means excluding invalid
characters like "/" or "$" or "%". Spaces are nasty too.

Since we've only ever made event IDs that match this pattern anyway,
I figure we should just enforce it as a rule.

https://github.com/burningmantech/ranger-ims-server/issues/7