add REPORT_ENTRY.ATTACHED_FILE, fix foreign key bug (#1619)
There was a lingering problem with FIELD_REPORT__REPORT_ENTRY, in that
one of its constraints referenced the now-renamed INCIDENT_REPORT table
(now FIELD_REPORT). Suddenly I was starting to see problems with that
constraint locally, which is how I noticed it never got updated in
MariaDB. This migration removes possible old forms of the constraint,
and adds a new consistent one instead.

Also, this adds the new ATTACHED_FILE column, which is what I was really
working on. The foreign key bug came about all of a sudden on my
computer, which was really weird.