add initial support for file attachments on incidents
I'll write up more on the issue, #7

This adds "working" support for adding attachments to incidents.
The files can be uploaded/served either via local filesystem or
S3, as configured.

Still to do:
* Try to get the Twisted server not to block other requests while
  an attachment is being uploaded or downloaded. It sucks that
  the single-threadedness of it means currently that other users of
  IMS will see unresponsiveness while some user is uploading/downloading
  a large file.
* Add support for attachments to Field Reports as well.
* Find a way to block excessively large requests from the server side.
  I've added a client-side check, but still, if a massive request comes
  in (multigig), then the IMS server totally halts indefinitely, since
  the server runs out of memory and thrashes until it's restarted.

I don't love where this is at the moment, and I don't consider it ready
for BRC 2025. I might want to spend some time in the fall/winter next
year to get this right.