simplify EventSource client code
this cleans up some Promise usage, simplifies a setTimeout, internalizes
the fire-and-forget Promise to requestEventSourceLock (less of a
footgun), and removes "useCapture = true" setting for some of the
EventSource addEventListener calls. That didn't seem to do anything,
which is understandable, since capture/bubble phases aren't really
relevant on EventSource events.