change the PK on EVENT_ACCESS to stop transaction blockages (#1658)
as detailed in 11-from-10.mysql, the previous PK was a problem
for automated testing, since attempts to change EVENT_ACCESSes
on different events at the same time would often fail and get
rolled back. We didn't really need "expression" as part of the
PK anyway, since things behave fine even if there is a duplicate
(event, expression) entry.

also this PR updates the playwright tests, which are now much
more successsful based on this DB schema tweak.

also this makes a minor tweak to the login page, so that we
autofocus on the data entry.