clean up the dark/light mode code
This moves that code into a separate JS file (necessary for moving
to JS modules) and it cleans up the invocation points for the theme
code. i.e. we now call immediately on <src> load to apply any stored
or preferred theme (to prevent flickering if done later), then we
set up the navbar theme selector after DOMContentLoaded.